Updated FavoriteSpecifications Class

+++ src/main/java/com/example/PokemonAPI/repository/FavoriteSpecifications.java
@@ -10,7 +10,7 @@ public class FavoriteSpecifications {
                 return criteriaBuilder.conjunction();
             }
 
-            return criteriaBuilder.equal(root.get("id").get("userId"), Long.parseLong(id));
+            return criteriaBuilder.equal(root.get("id").get("userId"), Integer.parseInt(id));
         };
     }
 }
